+++ application/views/app/adminhome.php
@@ -21,7 +21,7 @@
 </div>
 <div class="buttonOutline">
 	<a class="button" href="edit/<?= $post->id; ?>">Edit Post</a><br>
-	<a class="button" href="delete/<?= $post->id; ?>">Delete Post</a>
+	<a onclick="return confirm('Are you sure?');" class="button" href="delete/<?= $post->id; ?>">Delete Post</a>
 </div>
 	</div>
 <?php endforeach ?>
